警告消し 堀

+++ FarbeFahrt/FarbeFahrt/Root/Scene/BackLog.cpp
@@ -65,7 +65,8 @@ void BackLog::draw(Renderer& renderer)
 
 	for (int i = 0; i < drawable; ++i)
 	{
-		renderer.drawFontCenter(BackLogManager::GetRow(i + m_currentRow), outside.y + padding.y * inside.y + i * 48);
+		renderer.drawFontCenter(BackLogManager::GetRow(i + m_currentRow),
+			static_cast<int>(outside.y + padding.y * inside.y + i * 48));
 	}
 }
