+++ FarbeFahrt/FarbeFahrt/Root/Scene/Title.cpp
@@ -93,7 +93,8 @@ void Title::draw(Renderer& renderer)
 	if (System::GetWindowWidth() >= 1024)
 	{
 		renderer.drawTexture("TitleBigBackGround", Renderer::AspectType::LetterBox);
-		renderer.drawTexture(m_alphaSwitch ? "TitleBigBubble" : "TitleBigBubble2", Renderer::AspectType::LetterBox, t * 255.0f);
+		renderer.drawTexture(m_alphaSwitch ? "TitleBigBubble" : "TitleBigBubble2", Renderer::AspectType::LetterBox,
+			static_cast<int>(t * 255.0f));
 		renderer.drawTexture("TitleBigIvy", Renderer::AspectType::LetterBox);
 		renderer.drawTexture("TitleBigOther", Renderer::AspectType::LetterBox);
 	}
